+++ eslint-plugin-prettier.js
@@ -13,6 +13,7 @@
  * @typedef {import('prettier').FileInfoOptions} FileInfoOptions
  * @typedef {import('prettier').Options} PrettierOptions
  * @typedef {PrettierOptions & { onDiskFilepath: string, parserMeta?: ObjectMetaProperties['meta'], parserPath?: string, usePrettierrc?: boolean }} Options
+ * @typedef {(source: string, options: Options, fileInfoOptions: FileInfoOptions) => string} PrettierFormat
  */
 
 'use strict';
@@ -39,7 +40,7 @@ const { INSERT, DELETE, REPLACE } = generateDifferences;
 
 // Lazily-loaded Prettier.
 /**
- * @type {(source: string, options: Options, fileInfoOptions: FileInfoOptions) => string}
+ * @type {PrettierFormat}
  */
 let prettierFormat;
 
@@ -160,11 +161,11 @@ const eslintPluginPrettier = {
         const source = sourceCode.text;
 
         return {
-          Program() {
+          Program(node) {
             if (!prettierFormat) {
               // Prettier is expensive to load, so only load it if needed.
-              prettierFormat = require('synckit').createSyncFn(
-                require.resolve('./worker'),
+              prettierFormat = /** @type {PrettierFormat} */ (
+                require('synckit').createSyncFn(require.resolve('./worker'))
               );
             }
 
@@ -213,7 +214,7 @@ const eslintPluginPrettier = {
               let message = 'Parsing error: ' + err.message;
 
               const error =
-                /** @type {SyntaxError & {codeFrame: string; loc: SourceLocation}} */ (
+                /** @type {SyntaxError & {codeFrame: string; loc?: SourceLocation}} */ (
                   err
                 );
 
@@ -226,9 +227,10 @@ const eslintPluginPrettier = {
               }
               if (error.loc) {
                 message = message.replace(/ \(\d+:\d+\)$/, '');
+                context.report({ message, loc: error.loc });
+              } else {
+                context.report({ message, node });
               }
-
-              context.report({ message, loc: error.loc });
 
               return;
             }