+++ recommended.js
@@ -1,9 +1,10 @@
 const eslintConfigPrettier = require('eslint-config-prettier');
 const eslintPluginPrettier = require('./eslint-plugin-prettier');
 
-// Merge the contents of eslint-config-prettier into every
+// Merge the contents of eslint-config-prettier into config
 module.exports = {
   ...eslintConfigPrettier,
+  name: 'eslint-plugin-prettier/recommended',
   plugins: {
     ...eslintConfigPrettier.plugins,
     prettier: eslintPluginPrettier,